fix: eslint config 수정
