--- apps/web/.eslintrc.json
@@ -1,3 +1,0 @@
-{
-  "extends": "@repo/eslint-config/next.js"
-}
