+++ apps/web/src/app/page.tsx
@@ -1,3 +1,3 @@
 export default function Home() {
-  return <div>웹 1팀 파이팅</div>;
+  return <div>웹 1팀 파이팅!</div>;
 }
