+++ eslint.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  root: true,
+  extends: ['@repo/eslint-config/base.mjs'],
+};
