+++ packages/eslint-config/package.json
@@ -3,11 +3,9 @@
   "version": "0.0.0",
   "type": "module",
   "private": true,
-  "exports": {
-    "./base": "./base.js",
-    "./next-js": "./next.js",
-    "./react-internal": "./react-internal.js"
-  },
+  "files": [
+    "next.mjs"
+  ],
   "devDependencies": {
     "@eslint/js": "^9.17.0",
     "@next/eslint-plugin-next": "^15.1.0",
